Structured MVVM on update_profile_page_view
