+++ lib/locator.dart
@@ -2,6 +2,7 @@
 import 'package:get_it/get_it.dart';
 import 'package:talawa/utils/gql_client.dart';
 import 'package:talawa/view_models/page_view_model/join_organization_viewModel.dart';
+import 'package:talawa/view_models/page_view_model/update_profile_page_viewModel.dart';
 import 'package:talawa/view_models/swtich_org_vm.dart';
 import 'services/api_.dart';
 
@@ -10,7 +11,10 @@ GetIt locator = GetIt.instance;
 void setupLocator() {
   locator.registerLazySingleton(() => API());
   locator.registerFactory(() => GraphQLConfiguration());
+
+  //Register ViewModels
   locator.registerFactory(() => JoinOrgnizationViewModel());
+  locator.registerFactory(() => UpdateProfilePageViewModel());
 
   locator.registerFactory(() => SwitchOrgModel());
 }
