+++ lib/views/pages/_pages.dart
@@ -1,4 +1,4 @@
 export 'home_page.dart';
 export 'login_signup/login_page.dart';
-export 'organization/join_organization.dart';
+export 'organization/join_organization_view.dart';
 export 'organization/profile_page.dart';
