+++ test/widget_tests/update_profile_widget_test.dart
@@ -8,7 +8,7 @@ import 'package:talawa/services/post_provider.dart';
 import 'package:talawa/services/preferences.dart';
 import 'package:talawa/utils/gql_client.dart';
 import 'package:talawa/utils/ui_scaling.dart';
-import 'package:talawa/views/pages/organization/update_profile_page.dart';
+import 'package:talawa/views/pages/organization/update_profile_page_view.dart';
 
 Widget createProfileUpdateScreen() => MultiProvider(
       providers: [
@@ -23,7 +23,7 @@ Widget createProfileUpdateScreen() => MultiProvider(
       child: MaterialApp(
         home: Builder(builder: (BuildContext context) {
           SizeConfig().init(context);
-          return const UpdateProfilePage(
+          return const UpdateProfileView(
             userDetails: [
               {
                 "firstName": "Test",
